requirements.txt
Django==1.11.4
djangorestframework==3.6.4
pyjwt==1.6.4
django-session-timeout==0.0.3 
pycrypto==2.6.1

